add flag

+++ build_linux.sh
@@ -2,34 +2,42 @@
 
 export PING_SLEEP=30s
 export WORKDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-export BUILD_OUTPUT=$WORKDIR/build.out
 
-touch $BUILD_OUTPUT
+if [ "$REDIRECT" = "ON" ] ; then
 
-dump_output() {
-   echo Tailing the last 500 lines of output:
-   tail -500 $BUILD_OUTPUT
-}
-error_handler() {
-  echo ERROR: An error was encountered with the build.
-  dump_output
-  exit 1
-}
-# If an error occurs, run our error handler to output a tail of the build
-trap 'error_handler' ERR
+    export BUILD_OUTPUT=$WORKDIR/build.out
+    touch $BUILD_OUTPUT
 
-# Set up a repeating loop to send some output to Travis.
+    dump_output() {
+    echo Tailing the last 500 lines of output:
+    tail -500 $BUILD_OUTPUT
+    }
+    error_handler() {
+    echo ERROR: An error was encountered with the build.
+    dump_output
+    exit 1
+    }
+    # If an error occurs, run our error handler to output a tail of the build
+    trap 'error_handler' ERR
 
-bash -c "while true; do echo \$(date) - building ...; sleep $PING_SLEEP; done" &
-PING_LOOP_PID=$!
+    # Set up a repeating loop to send some output to Travis.
+
+    bash -c "while true; do echo \$(date) - building ...; sleep $PING_SLEEP; done" &
+    PING_LOOP_PID=$!
+fi
 
 export num_cores=`grep -c processor /proc/cpuinfo`
 
 cd OpenBLAS
 
 # make clean
-make -j${num_cores} >> $BUILD_OUTPUT 2>&1
-make install PREFIX=$PWD >> $BUILD_OUTPUT 2>&1
+if [ "$REDIRECT" = "ON" ] ; then
+    make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+    make install PREFIX=$PWD >> $BUILD_OUTPUT 2>&1
+else
+    make -j${num_cores}
+    make install PREFIX=$PWD
+fi
 
 cd ../Ipopt
 
@@ -59,10 +67,17 @@ cd ../Metis
 
 cd ../..
 
-./configure --prefix=$PWD --with-blas="-L$PWD/../OpenBLAS/lib -lopenblas" --with-lapack="-L$PWD/../OpenBLAS/lib -lopenblas" --with-mumps=no --with-asl=no --with-pardiso=no ADD_CFLAGS=-fopenmp ADD_FFLAGS=-fopenmp ADD_CXXFLAGS=-fopenmp >> $BUILD_OUTPUT 2>&1
-make clean >> $BUILD_OUTPUT 2>&1
-make -j${num_cores} >> $BUILD_OUTPUT 2>&1
-make install >> $BUILD_OUTPUT 2>&1
+if [ "$REDIRECT" = "ON" ] ; then
+    ./configure --prefix=$PWD --with-blas="-L$PWD/../OpenBLAS/lib -lopenblas" --with-lapack="-L$PWD/../OpenBLAS/lib -lopenblas" --with-mumps=no --with-asl=no --with-pardiso=no ADD_CFLAGS=-fopenmp ADD_FFLAGS=-fopenmp ADD_CXXFLAGS=-fopenmp >> $BUILD_OUTPUT 2>&1
+    make clean >> $BUILD_OUTPUT 2>&1
+    make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+    make install >> $BUILD_OUTPUT 2>&1
+else
+    ./configure --prefix=$PWD --with-blas="-L$PWD/../OpenBLAS/lib -lopenblas" --with-lapack="-L$PWD/../OpenBLAS/lib -lopenblas" --with-mumps=no --with-asl=no --with-pardiso=no ADD_CFLAGS=-fopenmp ADD_FFLAGS=-fopenmp ADD_CXXFLAGS=-fopenmp
+    make clean
+    make -j${num_cores}
+    make install
+fi
 
 if [ $? -ne 0 ] ; then
     exit 1
@@ -71,19 +86,34 @@ fi
 cd ../alglib
 mkdir build
 cd build
-cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. .. >> $BUILD_OUTPUT 2>&1
-make clean >> $BUILD_OUTPUT 2>&1
-make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+
+if [ "$REDIRECT" = "ON" ] ; then
+    cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. .. >> $BUILD_OUTPUT 2>&1
+    make clean >> $BUILD_OUTPUT 2>&1
+    make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+else
+    cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. ..
+    make clean
+    make -j${num_cores}
+fi
 
 if [ $? -ne 0 ] ; then
     exit 1
 fi
 
 cd ../../clp
-./configure --prefix=$PWD >> $BUILD_OUTPUT 2>&1
-make clean >> $BUILD_OUTPUT 2>&1
-make -j${num_cores} >> $BUILD_OUTPUT 2>&1
-make install >> $BUILD_OUTPUT 2>&1
+
+if [ "$REDIRECT" = "ON" ] ; then
+    ./configure --prefix=$PWD >> $BUILD_OUTPUT 2>&1
+    make clean >> $BUILD_OUTPUT 2>&1
+    make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+    make install >> $BUILD_OUTPUT 2>&1
+else
+    ./configure --prefix=$PWD
+    make clean
+    make -j${num_cores}
+    make install
+fi
 
 if [ $? -ne 0 ] ; then
     exit 1
@@ -94,16 +124,28 @@ mkdir build
 cd build
 
 if [ "$BUILD_TEST" = "ON" ] ; then
-    cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. -DTEST=ON .. >> $BUILD_OUTPUT 2>&1
-    make clean >> $BUILD_OUTPUT 2>&1
-    make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+    if [ "$REDIRECT" = "ON" ] ; then
+        cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. -DTEST=ON .. >> $BUILD_OUTPUT 2>&1
+        make clean >> $BUILD_OUTPUT 2>&1
+        make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+    else
+        cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. -DTEST=ON ..
+        make clean
+        make -j${num_cores}
+    fi
 
     cd ../bin
     ./test_suite
 else
-    cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. .. >> $BUILD_OUTPUT 2>&1
-    make clean >> $BUILD_OUTPUT 2>&1
-    make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+    if [ "$REDIRECT" = "ON" ] ; then
+        cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. .. >> $BUILD_OUTPUT 2>&1
+        make clean >> $BUILD_OUTPUT 2>&1
+        make -j${num_cores} >> $BUILD_OUTPUT 2>&1
+    else
+        cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$PWD/.. ..
+        make clean
+        make -j${num_cores}
+    fi
 
     if [ $? -ne 0 ] ; then
         exit 1
@@ -154,7 +196,11 @@ cd Ipopt/contrib/JavaInterface
 export CLASSPATH=$PWD:$CLASSPATH
 export CPLUS_INCLUDE_PATH=$JAVA_HOME/include/linux
 
-make >> $BUILD_OUTPUT 2>&1
+if [ "$REDIRECT" = "ON" ] ; then
+    make >> $BUILD_OUTPUT 2>&1
+else
+    make
+fi
 
 if [ $? -ne 0 ] ; then
     exit 1
@@ -162,8 +208,11 @@ fi
 
 echo "Portfolio - Optimizer building completed!"
 
-# The build finished without returning an error so dump a tail of the output
-dump_output
 
-# nicely terminate the ping output loop
-kill $PING_LOOP_PID
+if [ "$REDIRECT" = "ON" ] ; then
+    # The build finished without returning an error so dump a tail of the output
+    dump_output
+
+    # nicely terminate the ping output loop
+    kill $PING_LOOP_PID
+fi
